controller/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.27.2
openai==1.58.1
python-dotenv==1.0.1
